Use Fuel as default http client, to support Android SDK 19+

+++ src/main/java/com/aurora/gplayapi/data/models/App.kt
@@ -56,7 +56,7 @@ class App(var packageName: String) {
     var testingProgramAvailable = false
     var testingProgramOptedIn = false
     var offerType = 0
-    var versionCode: Long = 0
+    var versionCode: Int = 0
     var installs: Long = 0
     var size: Long = 0
     var rating: Rating = Rating()
